requirements.txt
flake8==7.2.0
matplotlib==3.10.1
numpy==1.23.5
pandas==2.2.3
plotly==6.0.1
pylint==3.3.6
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
seaborn==0.13.2
bandit==1.8.3
mypy==1.15.0
fastapi==0.115.12
uvicorn==0.34.1
pandas-stubs
types-PyYAML